refactoring navbar listeners
